refactor: unify naming for transfer progress and step components
- Renamed sender and receiver transfer progress components to `TransferProgress` for consistency.
- Standardized step component names from `WaitingForSender/Recipient` to `WaitForSender/Receiver` for improved clarity and brevity.
- Updated all related imports, exports, and usages across the codebase.
- No functional changes; improves naming consistency and code maintainability.
