+++ relayr-ui/src/components/file-transfer/receiver/steps/Step3_Receiving.tsx
@@ -10,7 +10,7 @@ import { MotionButton } from "@/components/animations/motion-button";
 import { TextShimmer } from "@/components/motion-primitives/text-shimmer";
 
 // Internal Components
-import { ReceiverTransferProgress } from "../components";
+import { TransferProgress } from "../components";
 import { TransferFileCard, TransferHeader } from "../../shared";
 
 // Animation Variants
@@ -89,7 +89,7 @@ export default function Step3_Receiving() {
         <TransferFileCard fileMetadata={fileMetadata} className="mt-2" />
       </motion.div>
 
-      <ReceiverTransferProgress />
+      <TransferProgress />
 
       <motion.div className="w-full flex flex-col space-y-3 mt-2">
         <TextShimmer className="text-center" duration={1}>
