+++ relayr-ui/src/components/file-transfer/sender/steps/Step5_Sending.tsx
@@ -11,7 +11,7 @@ import { TextShimmer } from "@/components/motion-primitives/text-shimmer";
 
 // Internal Components
 import ShareableLinkInput from "@/components/ShareableLinkInput";
-import { SenderTransferProgress } from "../components";
+import { TransferProgress } from "../components";
 import { TransferFileCard, TransferHeader } from "../../shared";
 
 // Animation Variants
@@ -96,7 +96,7 @@ export default function Step5_Sending() {
         <TransferFileCard fileMetadata={fileMetadata} />
       </motion.div>
 
-      <SenderTransferProgress />
+      <TransferProgress />
 
       <motion.div
         variants={fileListItemVariants}
